observatorio_banda_larga_fixa/requirements.txt
pyspark>=3.5.0
requests>=2.32.3
python-dotenv>=1.0.1
pandas>=2.2.2
pyarrow>=16.1.0
